feat: implement broadcast channel context
