--- config/nebula.pkl
@@ -1,61 +1,0 @@
-amends "https://pkl.sh/UIUX.pkl";
-
-home = new Home {
-	hero = new Hero {
-		titleFirst = "Simplified";
-		titleEmphasis = "self-custody";
-		titleSecond = "for everyone";
-		subtitle = "Sonr is a modern re-imagination of online user identity, empowering users to take ownership of their digital footprint and unlocking a new era of self-sovereignty.";
-		primaryButton = new Button {
-			text = "Get Started";
-			href = "/register";
-		};
-		secondaryButton = new Button {
-			text = "Learn More";
-			href = "/about";
-		};
-		image = new Image {
-			src = "https://cdn.sonr.id/img/hero-clipped.svg";
-			width = "500";
-			height = "500";
-		};
-		stats {
-			new Stat {
-				value = "476K";
-				label = "Assets packed with power beyond your imagination.";
-			};
-			new Stat {
-				value = "1.44K";
-				label = "Assets packed with power beyond your imagination.";
-			};
-			new Stat {
-				value = "1.5M+";
-				label = "Assets packed with power beyond your imagination.";
-			}
-		};
-	};
-};
-
-register = new ModalForm {
-	title = "Register";
-	description = "Register your Sonr account to start using the platform.";
-	id = "register-start";
-};
-
-login = new ModalForm {
-	title = "Login";
-	description = "Login to your Sonr account to start using the platform.";
-	id = "login-start";
-};
-
-authorize = new ModalForm {
-	title = "Authorize";
-	description = "Authorize your Sonr account to start using the platform.";
-	id = "authorize-start";
-};
-
-privacyConsent = new ModalForm {
-	title = "Privacy Consent";
-	description = "Privacy Consent";
-	id = "privacy-consent-start";
-};
