+++ internal/ctx/swt/macaroon.go
